add Downloder.get API & Request, Response objects
The Response object exposes the following info:

- url: the URL that was ultimately requested
- status: the HTTP status code of the response
- response: the HTTP response line
- headers: response headers

The headers are returned with their keys lowercased. Doing this now is
future-proofing any users of this since HTTP/2 headers are always
lowercase and many users are probably incorrectly checking for case-
sensitive headers like "Content-Type" which are going to be returned
as "content-type" from an HTTP/2 server. We may as well get ahead of
the curve on this one since we have no users yet.