fix laziness from libcurl/libuv example code
In https://curl.haxx.se/libcurl/c/multi-uv.html the defintion of
start_timeout has the following code with a suspicious comment:

    if(timeout_ms == 0)
      timeout_ms = 1; /* 0 means directly call socket_action,
                         but we'll do it in a bit */

It turns out that if libcurl fetches a URL and the fetch never needs
to yield to the scheduler, this causes the download to hang until
some other event (like a key press) triggers the scheduler. This was
discovered while implementing and testing proper error handling. The
case where this hangs is a URL where libcurl knows before even trying
to talk to a server that the download will fail, such as one with an
invalid or unsupported protocol. In such a case, without this fix,
the download call would hang until you press a key or something else
happens to trigger a libuv event.