fix broken `download(url, path; headers)` method
This method was broken because of a trivial typo: `,` versus `;` in
the method signature, which escaped notice because the tests were not
sufficiently thorough. I could just fix the typo and add some more
tests, but since (a) this is exactly the kind of thing ArgTools is for
and (b) the Tar stdlib already depends on ArgTools, we might as well
just use it and use its test utilities for more thorough testing.