Multi: add currently-empty `set_defaults` method for multi-sockets
For a while, I was setting the max host connections to 50 because I
was seeing failures in the concurrent tests when making 100 concurrent
connections (anything more than around 65). However, I've since learned
that this can be prevented by calling yield inside of a lock before
starting each download (see later commits), so since that's already
happening at this point, we don't need to set a maximum on connections.
I'm leaving the `set_defaults` method in the code, however, so that if
we want to set defaults on `Multi` objects in the future, we can.