timer_callback: no functional change
This doesn't change behavior at all but makes it easier to go back
and forth between doing the meat of the `timeout_callback` function
immediately when `timeout_ms == 0` or deferring it until later by
setting a one millisecond timer.

After experimenting with both approaches, it seems that not handling
the `timeout_ms == 0` case immediately can cause hangs when handling
error cases like invalid domains and other kinds of failures. This
only seems to occur when running the tests repeatedly. My guess is
that if the system does a full DNS query then there is a delay so
that case doesn't get hit but if it has already done a failed DNS
query recently, there's no delay in getting the error response and
we hit the zero timeout case and setting a short timeout somehow
causes a failure to catch the event at all, leading to a hang.