thread & memory safety: lock & preserve Multi objects
There are three safety & correctness measures added in this commit:

1. Add a lock to each `Multi` object and acquire the lock before
any action that modifies the state of the object or its handle.

2. Use `preserve_handle` to ensure that each `Multi` object is not
garbage collected while any downloads it is responsible for are in
progress. This is done at a fairly coarse-grained level: if any of
the downloads are still ongoing, we prevent collection. We also
keep a list of all the `Easy` handles that each `Multi` handle is
processing so that those are rooted as well, preventing them from
being garbage collected while they are in use. This is probably
overkill since an `Easy` handle cannot be collected in the middle
of the `download` function since we call `done!` on it at the end
but it doesn't cost much to keep this list and it seems useful to
have this information for reflection purposes anyway.

3. It addresses one other issueby adding a global DOWNLOAD lock and
calling `yield()` while the lock is held. This allows any downloads
already in progress to process events before starting any new ones.
Just calling `yield()` without the lock doesn't fix this issue
since if there are new downloads that are trying to start it will
allow them to start first, piling on even more work. Locking first
prevents that from happening since any additional downloads have to
wait for the lock, so the `yield()` actually allows older downloads
to finish. Without this, highly concurrent downloads, as in the
tests, can cause "cound not connect to server" errors, that I
suspect are due to events not being processed fast enough.