API: fully document `download`; official `request` API (#46)
In addition to making the progress API official, this tweaks it: since we are explicitly downloading (not uploading), the callback only gets the download values: `dl_total` and `dl_now`. This also documents the `Downloader` type.

Most importantly, it adds an official request API returning a `Response` object or throwing a `RequestError` if there's a libcurl error. The `download` function also throws a `RequestError` on libcurl error or on non-2xx status code.