progress: drop legacy API; allow 2-args p-func to request when clear
The most questionable part of this is that if `request` is called
with a 2-arg progress function, I'm adding the download and upload
values to give a summed total/now pair.

This also turns off progress callbacks hooks altogether if no
progress callback is passed by the user. This should have no visible
effect; there's just no point in having libcurl call progress hook
if we know that the hook doesn't do anything.