request with no `output` defaults to `HEAD` method
We were already doing a `PUT` request if there is `input` to the
request, this extends that to do `GET` if there is `output` but
`HEAD` if there is no `input` or `output`. The implementation is
a bit sutbler than that: we use `libcurl`'s `CURLOPT_NOBODY` and
`CURLOPT_UPLOAD` options, which incidentally switches the default
method that it uses for HTTP/S requests and has corresponding
appropriate effects on uploads/downloads by other protocols.

This also changes the default for `input` and `output` to `nothing`
instead of using `devnull` as a sentinel, which had the unfortunate
side-effect that if explicitly pass `devnull` as an input or output
argument, intending to make the same request but provide an empty
input or ignore any output, it would change the behavior of the
request. This was already present if you tried to do a `PUT` with
an empty request body: using `devnull` turned it into a `GET` but
using `"/dev/null"` or some other empty file didn't. With `nothing`
as a sentinel value, you can safely pass `devnull` as an input or
and output and have it work the same as any other value but empty.