add connect timeout; reinstate speed limit
- This adds a connection timeout of 30 seconds, closing #125
- It re-instates the lower speed limit that was briefly added in #85 and
  was accidentally reverted in the following pull request (#86)

In combination, this should prevent downloads from hanging, whether
because the server is unavailable and so the connection is never
established, or because the server never sends any data after the
connection is established. This does not put an absolute timeout on
downloads since a very large download can take arbitrarily long while
still making progress.