fix tests for httpbingo upgrade (#152)
This test started failing because httpbingo now serves a 300 response
with a location header, which gets followed so the response object ends
up being a 200 OK instead of the expected 300. This changes the test to
use 200, 400 and 404, which should never redirect.

(cherry picked from commit dbb062584961888c6abddcb186d0248bec7763b8)