error message: include string `RequestError`
This is done in the example in `Base.showerror`: https://github.com/JuliaLang/julia/blob/2049baaf31edc75c4403177e21bb1c3be49fb683/base/errorshow.jl#L16

It would've been helpful for https://github.com/JuliaCloud/AWS.jl/issues/541 to understand what kind of error it is. E.g. if it's a RequestError, then it's already being retried and we must've hit the retry limit; if it wasn't a RequestError, it was likely uncaught and therefore we'd be missing a retry.