add concurrent connection semaphore to prevent DNS resolve races
The libcurl spawns a new thread for each DNS resolution that it needs to
do, which can cause problems when too many connections are started at
the same time. This adds a semaphore which is acqured when a connection
starts and gets released once the actual request starts. This limits the
number of concurrent requests that are in the connecting state to the
size of the semaphore (16). Hopefully this prevents the "getaddrinfo()
thread failed" errors that occur sometimes.