add try/catch to all callback functions (never unwind through them)
This may help https://github.com/JuliaLang/Downloads.jl/issues/196. The
issue there may be that callback functions raise Julia errors sometimes
(they shouldn't but stuff happens, ya know?). This wraps each one in a
try/catch block and logs an "unkonwn error" message if that happens and
returns from the callback with an appropriate return code to indicate
error. The libcurl docs aren't super clear on these and this is hard to
test so hopefully I got it right, but we'll see. Should be better.